pygreen-bd/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
distlib==0.3.8
filelock==3.13.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector==2.2.9
mysql-connector-python==9.4.0
pillow==11.1.0
platformdirs==4.1.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
Werkzeug==3.1.3
